:recycle: Remove old IO comm classes and move NetworkTableWrapper to lib
